requirements.txt
certifi==2018.1.18
chardet==3.0.4
ChatterBot==0.8.4
chatterbot-corpus==1.1.1
click==6.7
Flask==0.12.2
future==0.16.0
idna==2.6
itsdangerous==0.24
Jinja2==2.10
MarkupSafe==1.0
mathparse==0.1.1
nltk==3.2.5
oauthlib==2.0.6
pymessenger==0.0.7.0
pymongo==3.6.0
python-dateutil==2.6.1
python-telegram-bot==9.0.0
python-twitter==3.4
PyYAML==3.12
redis==2.10.6
requests==2.18.4
requests-oauthlib==0.8.0
requests-toolbelt==0.8.0
six==1.11.0
SQLAlchemy==1.2.3
tweepy==3.5.0
urllib3==1.22
Werkzeug==0.14.1
